ananya-service/src/main/java/org/motechproject/ananya/service/JobAidService.java
package org.motechproject.ananya.service;

import org.motechproject.ananya.domain.FrontLineWorker;
import org.motechproject.ananya.request.JobAidPromptRequest;
import org.motechproject.ananya.response.JobAidCallerDataResponse;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Service;

@Service
public class JobAidService {

    private FrontLineWorkerService frontLineWorkerService;
    private OperatorService operatorService;

    @Autowired
    public JobAidService(FrontLineWorkerService frontLineWorkerService, OperatorService operatorService) {
        this.frontLineWorkerService = frontLineWorkerService;
        this.operatorService = operatorService;
    }

    public void updateJobAidPrompts(JobAidPromptRequest jobAidPromptRequest) {
        frontLineWorkerService.updatePromptsForFLW(
                jobAidPromptRequest.getCallerId(),
                jobAidPromptRequest.getPromptList());
    }

    public JobAidCallerDataResponse createCallerData(String msisdn, String operator) {
        FrontLineWorker frontLineWorker = frontLineWorkerService.createNew(msisdn, operator);
        int currentJobAidUsage = frontLineWorker.getCurrentJobAidUsage();
        Integer allowedUsagePerMonthForOperator = operatorService.findMaximumUsageFor(operator);
        return new JobAidCallerDataResponse(
                frontLineWorker.status().isRegistered(),
                currentJobAidUsage,
                allowedUsagePerMonthForOperator,
                frontLineWorker.getPromptsHeard());
    }
}
